fix log clear timeout
When clearing large number of event entries, the eraseAll function will
cost long time for removing persistent files on BMC, which will lead to
method call timeout from bmcweb. This patch moves the event entry files
operation to child process to resolve the issue.

```
Tested:
1. Created 6000 logs and then using event log clear redfish to clear the
logs. No error return after change and the entries removed from disk.

$curl -v -w "%{time_total}\n" -o response.txt -k -H 'Content-Type: application/json' -X POST https://bmc-ip/redfish/v1/Systems/Bluefield/LogServices/EventLog/Actions/LogService.ClearLog
  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0*   Trying 10.237.23.75:443...
...
< HTTP/1.1 204 No Content
...
* Connection #0 to host ldev-platform-14-061-bmc left intact
4.845050

root@dpu-bmc:~# ls /var/lib/logging/logs/errors/
1        Manager/ SEL/
root@dpu-bmc:~# ls /var/lib/logging/logs/errors/SEL/
2

2. Reboot BMC during log clear, the logs and disk files will be removed
after BMC boot up.

3. Kill child process of phosphor-log-manager, the log files on disk
will be left on temporary folder, another call of event log clear
redfish request or BMC reboot will clear the temporary files.
```
Fixes jira https://redmine.mellanox.com/issues/4268225

+++ log_manager.cpp
@@ -831,6 +831,7 @@ void Manager::checkAndRemoveBlockingError(uint32_t entryId)
 
 size_t Manager::eraseAll()
 {
+#ifndef ENABLE_ERASE_WITH_MULTIPLE_PROCESS
     std::vector<uint32_t> logIDWithHwIsolation;
     for (auto& func : Extensions::getLogIDWithHwIsolationFunctions())
     {
@@ -892,10 +893,13 @@ size_t Manager::eraseAll()
             entryId = 0;
         }
     }
+#else
+        eraseAllInChildProcess(DEFAULT_BIN_NAME);
+#endif
     return entriesSize;
 }
 
-void Manager::erase(uint32_t entryId)
+void Manager::erase(uint32_t entryId, bool removePersistentFiles)
 {
     auto entryFound = entries.find(entryId);
 
@@ -930,19 +934,23 @@ void Manager::erase(uint32_t entryId)
             }
         }
 
-        if (!(binName.compare(DEFAULT_BIN_NAME) == 0))
+        if (removePersistentFiles)
         {
-            deletePath = std::string(phosphor::logging::paths::error()) + "/" + binName;
+            if (!(binName.compare(DEFAULT_BIN_NAME) == 0))
+            {
+                deletePath = std::string(phosphor::logging::paths::error()) + "/" + binName;
+            }
+
+            // lg2::info("Deleting Entry of Bin: {BIN_NAME}", "BIN_NAME",
+            // binName);
+            // lg2::info("Bin of Incoming Entry: {DELETE_PATH}",
+            // "DELETE_PATH", deletePath);
+
+            // Delete the persistent representation of this error.
+            fs::path errorPath(deletePath);
+            errorPath /= std::to_string(entryId);
+            fs::remove(errorPath);
         }
-
-        // lg2::info("Deleting Entry of Bin: {BIN_NAME}", "BIN_NAME", binName);
-        // lg2::info("Bin of Incoming Entry: {DELETE_PATH}", "DELETE_PATH",
-        //           deletePath);
-
-        // Delete the persistent representation of this error.
-        fs::path errorPath(deletePath);
-        errorPath /= std::to_string(entryId);
-        fs::remove(errorPath);
 
         auto removeId = [](std::set<uint32_t>& ids, uint32_t id) {
             auto it = std::find(ids.begin(), ids.end(), id);
@@ -1006,6 +1014,12 @@ void Manager::restore()
     {
         return;
     }
+
+#ifdef ENABLE_ERASE_WITH_MULTIPLE_PROCESS
+    // Check and remove the temporary files for deletion.
+    util::removeStagedForEraseEntries(DEFAULT_BIN_NAME, binNameMap);
+#endif
+
 
     // using recursive_directory_iterator to get every directory
     for (const auto& file : std::filesystem::recursive_directory_iterator(dir))
@@ -1185,6 +1199,42 @@ bool Manager::deleteAll(
         }
     }
 
+    return true;
+}
+
+bool Manager::deleteAllTypes(const std::string& nspace)
+{
+#ifndef ENABLE_ERASE_WITH_MULTIPLE_PROCESS
+    auto binPresent = false;
+    Bin* thisBin;
+    for (auto& pair : binNameMap)
+    {
+        if (pair.first == nspace)
+        {
+            binPresent = true;
+            thisBin = &(pair.second);
+            break;
+        }
+    }
+    // If bin is not present then return error
+    if (!binPresent)
+    {
+        throw sdbusplus::xyz::openbmc_project::Common::Error::
+            ResourceNotFound();
+    }
+    // Info Errors
+    while (getInfoErrSize(nspace) != 0)
+    {
+        erase(*(thisBin->infoEntries.begin()));
+    }
+    // Real Errors
+    while (getRealErrSize(nspace) != 0)
+    {
+        erase(*(thisBin->errorEntries.begin()));
+    }
+#else
+    eraseAllInChildProcess(nspace);
+#endif
     return true;
 }
